requirements.txt
alembic==1.18.1
blinker==1.9.0
cached-property==2.0.1
click==8.3.1
dotenv==0.9.9
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
psycopg2==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.46
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.15.0
Werkzeug==3.1.5
